Create README.md
# Pandemic Flu Spread 

This folder contains the code to develop the simulation solution to the  pandemic flu 
spread. Here is a brief description of each file:

- *pandemic_class.py* - Contains the code, specifically the classes, to create the
simulation of the pandemic. These classes are responsible for creating susceptible, 
  infected and recovered students as wells as to setup the simulation environment 
  which is the job of the <u>PandemicSim</u> class
  
- *main_findings.ipynb* - Jupyter notebook with some of the code and plots used in 
the main findings section in the report.
  
- *SIR_modelling.ipynb* - Jupyter notebook with the SIR model as an extent to the main
simulation approach that was followed in this project
  
- *sim_test.py* - Initial setup and testing of the simulation model and classes. 
Included only for reference.
  
- *plot_utils* - Useful function to plot results

- *plots* folder - Folder to store some plots to use in the report

- *base.py* - a file that contains code to call the root folder 

## Python libraries
We used Python 3.8 and the following libraries for this project:
- numpy
- pandas
- scipy
- matplotlib
- seaborn